add behat feature for this feature, move tags collector to content bundle
