+++ features/twig/cache_tags_collector.feature
@@ -1,0 +1,57 @@
+@disable-fixtures
+Feature: Working with cache blocks http cache tags collector
+  In order to allways have correct http cache tags with used on page articles
+  As a template implementator
+  I want to be able to have http response tags set when content is cached and not cached
+
+  Scenario: Getting next article from content list
+    Given the following Tenants:
+      | organization | name | subdomain | domain_name | enabled | default |
+      | Default      | test |           | localhost   | true    | true    |
+
+    Given the following Content Lists:
+      | name                | type   |
+      | test content list   | manual |
+
+    Given the following Articles:
+      | title               | route      | status    |
+      | First Test Article  | Test Route | published |
+      | Second Test Article | Test Route | published |
+      | Third Test Article  | Test Route | published |
+
+    Given the following Content List Items:
+      | content_list      | article             |
+      | test content list | First Test Article  |
+      | test content list | Second Test Article |
+      | test content list | Third Test Article  |
+
+    And I render a template with content:
+     """
+      {% cache 'v1' {time: 10} %}
+        {% gimmelist article from articles %}
+            {{ article.title }}-{{ article.id }}
+        {% endgimmelist %}
+      {% endcache %}
+     """
+    Then rendered template should contain "First Test Article-1"
+    Then rendered template should contain "Second Test Article-2"
+    Then rendered template should contain "Third Test Article-3"
+    Then CacheBlockTagsCollector should have tag "article-1"
+    Then CacheBlockTagsCollector should have tag "article-2"
+    Then CacheBlockTagsCollector should have tag "article-3"
+
+    And I render a template with content:
+     """
+      {% cache 'v1' {time: 10} %}
+        {% gimmelist article from articles %}
+            // remove id's but it should still pass ass cached version should be used
+            {{ article.title }}
+        {% endgimmelist %}
+      {% endcache %}
+     """
+    Then rendered template should contain "First Test Article-1"
+    Then rendered template should contain "Second Test Article-2"
+    Then rendered template should contain "Third Test Article-3"
+    Then CacheBlockTagsCollector should have tag "article-1"
+    Then CacheBlockTagsCollector should have tag "article-2"
+    Then CacheBlockTagsCollector should have tag "article-3"
