+++ src/SWP/Behat/Contexts/TemplatingContext.php
@@ -7,6 +7,8 @@ namespace SWP\Behat\Contexts;
 use Behat\Behat\Context\Context;
 use Behat\Gherkin\Node\PyStringNode;
 use SWP\Bundle\ContentBundle\Loader\ArticleLoader;
+use SWP\Bundle\ContentBundle\Twig\Cache\CacheBlockTagsCollectorInterface;
+use Twig\Environment;
 
 final class TemplatingContext implements Context
 {
@@ -14,21 +16,24 @@ final class TemplatingContext implements Context
 
     private $articleLoader;
 
-    /**
-     * @var string
-     */
     private $lastRenderedContent;
 
-    public function __construct(\Twig_Environment $templating, ArticleLoader $articleLoader)
-    {
+    private $cacheBlockTagsCollector;
+
+    public function __construct(
+        Environment $templating,
+        ArticleLoader $articleLoader,
+        CacheBlockTagsCollectorInterface $cacheBlockTagsCollector
+    ) {
         $this->templating = $templating;
         $this->articleLoader = $articleLoader;
+        $this->cacheBlockTagsCollector = $cacheBlockTagsCollector;
     }
 
     /**
      * @Given I set :slug as a current article in the context
      */
-    public function IsetAsACurrentArticleInTheContext(string $slug): void
+    public function ISetAsACurrentArticleInTheContext(string $slug): void
     {
         $this->articleLoader->load('article', ['slug' => $slug]);
     }
@@ -69,6 +74,22 @@ final class TemplatingContext implements Context
     {
         if ($this->lastRenderedContent !== $templateContent->getRaw()) {
             throw new \Exception('The content is not equal!');
+        }
+    }
+
+    /**
+     * @Then CacheBlockTagsCollector should have tag :tagName
+     */
+    public function cacheblocktagscollectorShouldHaveTag(string $tagName)
+    {
+        if (!in_array($tagName, $this->cacheBlockTagsCollector->getCurrentCacheBlockTags())) {
+            throw new \Exception(
+                sprintf(
+                    'Tag %s was not found. Found tags: %s',
+                    $tagName,
+                    implode(', ', $this->cacheBlockTagsCollector->getCurrentCacheBlockTags())
+                )
+            );
         }
     }
 }
