+++ src/SWP/Bundle/ContentBundle/Resources/config/services.yml
@@ -16,6 +16,12 @@ services:
         autowire: true
         tags:
             - { name: swp.meta_loader.add }
+    SWP\Bundle\ContentBundle\Loader\ArticleLoader: '@swp_template_engine.loader.article'
+
+    SWP\Bundle\ContentBundle\Twig\Cache\CacheBlockTagsCollector:
+        autowire: true
+    SWP\Bundle\ContentBundle\Twig\Cache\CacheBlockTagsCollectorInterface: '@SWP\Bundle\ContentBundle\Twig\Cache\CacheBlockTagsCollector'
+
 
     swp_template_engine.loader.related_article:
         class: SWP\Bundle\ContentBundle\Loader\RelatedArticleLoader
@@ -349,4 +355,3 @@ services:
         public: true
 
     Doctrine\Common\Cache\Cache: '@doctrine_cache.providers.main_cache'
-    SWP\Bundle\ContentBundle\Loader\ArticleLoader: '@swp_template_engine.loader.article'
