+++ src/SWP/Bundle/CoreBundle/Resources/config/services.yml
@@ -266,15 +266,15 @@ services:
 
     SWP\Bundle\CoreBundle\Twig\Cache\Strategy\TagAwareIndexedChainingCacheStrategy:
         arguments:
-            - '@SWP\Bundle\CoreBundle\Twig\Cache\CacheBlockTagsCollector'
+            - '@SWP\Bundle\ContentBundle\Twig\Cache\CacheBlockTagsCollectorInterface'
             - '@fos_http_cache.http.symfony_response_tagger'
             -
                 gen: '@twig_cache.strategy.generational'
                 time: '@twig_cache.strategy.lifetime'
 
-    SWP\Bundle\CoreBundle\Twig\Cache\CacheBlockTagsCollector:
+    SWP\Bundle\ContentBundle\Twig\Cache\CacheBlockTagsCollector:
         autowire: true
-    SWP\Bundle\CoreBundle\Twig\Cache\CacheBlockTagsCollectorInterface: '@SWP\Bundle\CoreBundle\Twig\Cache\CacheBlockTagsCollector'
+    SWP\Bundle\ContentBundle\Twig\Cache\CacheBlockTagsCollectorInterface: '@SWP\Bundle\ContentBundle\Twig\Cache\CacheBlockTagsCollector'
 
     swp.facebook.service.instant_articles:
         class: SWP\Bundle\CoreBundle\Service\FacebookInstantArticlesService
